requirements.txt
numpy>=1.24,<3.0
pandas>=2.0,<3.0
yfinance>=0.2,<0.3
PyYAML>=6.0,<7.0
typer[all]>=0.12,<0.13
tqdm>=4.65,<5.0
